Update Node.js to latest 10.x.x version
Travis CI is currently defaulting to using Node.js v8.x.x because the
Travis config file does not specify to use Node at all. Autoprefixer-css
does not work with v8.x.x, so we need to use a newer version.

This commit adds a .nvmrc file specifying just a major version, and the
commands to read that .nvmrc file to the Travis configuration. Travis CI
will now use the whichever release with that major version it has
already cached.

In this commit the .nvmrc file specifies release 10.x.x, which is
the same as currently in tech-docs-gem[[1]].

[[1]]: https://github.com/alphagov/tech-docs-gem/blob/33f064e1f85d017e6a5926ec6c1b91c504621e02/.nvmrc